Updated HTML file with some important changes
I have added my name and Github link. Also, I have added 'using your Github account' in the of 'You can sign up anytime between October 1 and October 31'. 
The important change that I made was previously font was not applied to any text, you had used 'style:' and the font directly which was applicable in HTML 4 and not in HTML 5 so I changed it to 'style: "font-family: ' I applied font to the Heading as it was assigned earlier and added a font to Contributors. Also, I have changed the hyperlink color of the Github link of contributors. If you need any help contact me. 
Hope that helped : )